rlinf_dexhand/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rlinf_dexhand"
version = "0.1.0"
description = "Dexterous hand and data-glove drivers for RLinf"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "pyserial",
]

[project.optional-dependencies]
glove = ["pyyaml"]
aoyi = ["pymodbus==2.5.3"]
all = ["pyyaml", "pymodbus==2.5.3"]

[tool.setuptools.packages.find]
include = ["rlinf_dexhand*"]

[tool.setuptools.package-data]
"rlinf_dexhand.glove.psi_glove_driver" = ["*.yaml"]
